refactor(tests): rename test file so jest won't test it
- also consolidate all tests to single directory
