+++ __tests__/lib/babel.js
