+++ lib/jest.js
